meatbone slash/long status/short status/teleport 2
worst matches involve (out of 242):
               A Compress: 2.0661
             A Toxic Frog: 2.0661
          A Snake Carrier: 2.0661
                A Dispose: 2.0661
                 A Repair: 2.0661
                A Destroy: 2.0661
           A Midgar Swarm: 2.0661
                   C Bomb: 1.8079
                 C Reaper: 1.7531
          A Magic Barrier: 1.7062
                  C Porky: 1.5850
                 A Elixir: 1.4891
                  C Swine: 1.4646
                  C Ghoul: 1.4499
               C Skeleton: 1.4373
                      A 5: 1.4184
           C Floating Eye: 1.4067
                C Malboro: 1.3774
              C Wild Boar: 1.3774
          C King Behemoth: 1.3774

total: 4000
correct: 65.5%
time_outs: 0.3%
improvement: 31.1%
log loss: 0.668382

+++ simulator/src/sim/enums/condition.rs
@@ -219,7 +219,7 @@ impl Condition {
         CONDITION_NAMES[self.index()]
     }
 
-    pub fn condition_duration(self) -> Option<i8> {
+    pub fn condition_duration(self) -> Option<i32> {
         match self {
             Condition::Charm => Some(32),
             Condition::DontAct => Some(24),
